Improving SDK error/response handling (#19539)
* improving error/response handling

* updated request function to expose the response on error

* updated the composables where needed

* Create few-rules-talk.md

* ran prettier

* undid unintended type change

* added missing awaits for onResponse

* Update few-rules-talk.md

* Mark it as major change

* unpack directus errors instead of nesting

* ran prettier

* Update .changeset/few-rules-talk.md


---------
