+++ .changeset/few-rules-talk.md
@@ -1,0 +1,5 @@
+---
+"@directus/sdk": minor
+---
+
+Included the response object in thrown errors and added the request object in the `onResponse` hook
